Add support for detecting background size
The client script now looks for a white-ish rectangle. The
coordinates of the ball are scaled based on this rectangle to the
range (-1,1) in x and y. A red border is drawn around the rectangle

Also moves magic numbers to parameter variables at the top of the
script for easier modification.